rfc2136: merge Endpoints with same name/type into single Endpoint
