+++ provider/digitalocean/digital_ocean.go
@@ -112,38 +112,6 @@ func (p *DigitalOceanProvider) Zones(ctx context.Context) ([]godo.Domain, error)
 	return result, nil
 }
 
-// Merge Endpoints with the same Name and Type into a single endpoint with multiple Targets.
-func mergeEndpointsByNameType(endpoints []*endpoint.Endpoint) []*endpoint.Endpoint {
-	endpointsByNameType := map[string][]*endpoint.Endpoint{}
-
-	for _, e := range endpoints {
-		key := fmt.Sprintf("%s-%s", e.DNSName, e.RecordType)
-		endpointsByNameType[key] = append(endpointsByNameType[key], e)
-	}
-
-	// If no merge occurred, just return the existing endpoints.
-	if len(endpointsByNameType) == len(endpoints) {
-		return endpoints
-	}
-
-	// Otherwise, construct a new list of endpoints with the endpoints merged.
-	var result []*endpoint.Endpoint
-	for _, endpoints := range endpointsByNameType {
-		dnsName := endpoints[0].DNSName
-		recordType := endpoints[0].RecordType
-
-		targets := make([]string, len(endpoints))
-		for i, e := range endpoints {
-			targets[i] = e.Targets[0]
-		}
-
-		e := endpoint.NewEndpoint(dnsName, recordType, targets...)
-		result = append(result, e)
-	}
-
-	return result
-}
-
 // Records returns the list of records in a given zone.
 func (p *DigitalOceanProvider) Records(ctx context.Context) ([]*endpoint.Endpoint, error) {
 	zones, err := p.Zones(ctx)
@@ -176,7 +144,7 @@ func (p *DigitalOceanProvider) Records(ctx context.Context) ([]*endpoint.Endpoin
 
 	// Merge endpoints with the same name and type (e.g., multiple A records for a single
 	// DNS name) into one endpoint with multiple targets.
-	endpoints = mergeEndpointsByNameType(endpoints)
+	endpoints = endpoint.MergeEndpointsByNameType(endpoints)
 
 	// Log the endpoints that were found.
 	log.WithFields(log.Fields{
