+++ provider/digitalocean/digital_ocean_test.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 	"os"
 	"reflect"
-	"sort"
 	"testing"
 
 	"github.com/digitalocean/godo"
@@ -625,37 +624,4 @@ func TestDigitalOceanAllRecords(t *testing.T) {
 	if err == nil {
 		t.Errorf("expected to fail, %s", err)
 	}
-}
-
-func TestDigitalOceanMergeRecordsByNameType(t *testing.T) {
-	xs := []*endpoint.Endpoint{
-		endpoint.NewEndpoint("foo.example.com", "A", "1.2.3.4"),
-		endpoint.NewEndpoint("bar.example.com", "A", "1.2.3.4"),
-		endpoint.NewEndpoint("foo.example.com", "A", "5.6.7.8"),
-		endpoint.NewEndpoint("foo.example.com", "CNAME", "somewhere.out.there.com"),
-	}
-
-	merged := mergeEndpointsByNameType(xs)
-
-	assert.Equal(t, 3, len(merged))
-	sort.SliceStable(merged, func(i, j int) bool {
-		if merged[i].DNSName != merged[j].DNSName {
-			return merged[i].DNSName < merged[j].DNSName
-		}
-		return merged[i].RecordType < merged[j].RecordType
-	})
-	assert.Equal(t, "bar.example.com", merged[0].DNSName)
-	assert.Equal(t, "A", merged[0].RecordType)
-	assert.Equal(t, 1, len(merged[0].Targets))
-	assert.Equal(t, "1.2.3.4", merged[0].Targets[0])
-
-	assert.Equal(t, "foo.example.com", merged[1].DNSName)
-	assert.Equal(t, "A", merged[1].RecordType)
-	assert.Equal(t, 2, len(merged[1].Targets))
-	assert.ElementsMatch(t, []string{"1.2.3.4", "5.6.7.8"}, merged[1].Targets)
-
-	assert.Equal(t, "foo.example.com", merged[2].DNSName)
-	assert.Equal(t, "CNAME", merged[2].RecordType)
-	assert.Equal(t, 1, len(merged[2].Targets))
-	assert.Equal(t, "somewhere.out.there.com", merged[2].Targets[0])
 }
